face-blur-function/requirements.txt
fdk>=0.1.21
oci>=2.145.0
pillow>=9.0.0
python-dotenv>=1.0.0
opencv-python-headless>=4.8.0
numpy==1.26.4

